Basic spike IO and Threaded spike IO pass tests
This implements #2 and #1
