+++ src/CMakeLists.txt
@@ -50,9 +50,9 @@ if(${RAPIDJSON_FOUND})
 ## Include the TN Model JSON Parsing Library
 
 ## Set up the main project
-set(nemo_lib_srcs include/globals.h globals.cpp neuro/core.cpp neuro/core.h neuro/neuron_generic.cpp
-        neuro/neuron_generic.h  mapping.h mapping.cpp io/CoreOutput.cpp include/CoreOutput.h ../external/include/atomicops.h
-        ../external/include/readerwriterqueue.h)
+set(nemo_lib_srcs include/globals.h globals.cpp neuro/core.cpp neuro/core.h neuro/TrueNorthCore.h neuro/TrueNorthCore.cpp
+        neuro/neuron_generic.cpp neuro/neuron_generic.h  mapping.h mapping.cpp io/CoreOutput.cpp include/CoreOutput.h
+        ../external/include/atomicops.h ../external/include/concurrentqueue.h ../external/include/blockingconcurrentqueue.h)
 set(nemo_app_srcs main.cpp include/globals.h globals.cpp)
 
 ## Add required libraries
